requirements.txt
boto3==1.3.0
botocore==1.4.11
docutils==0.12
futures==3.0.5
invoke==0.12.2
jmespath==0.9.0
python-dateutil==2.5.2
PyYAML==3.11
six==1.10.0
